Test setup with Ava and SinonJS
