+++ package.json
@@ -4,7 +4,7 @@
   "description": "Only the dankest of dank memes, homies.",
   "main": "index.js",
   "scripts": {
-    "test": "make test",
+    "test": "ava",
     "lint": "eslint *.js"
   },
   "repository": {
@@ -34,12 +34,14 @@
     "gifsicle": "^4.0.1",
     "giphy-api": "^2.0.1",
     "isomorphic-fetch": "^2.2.1",
-    "opn": "^6.0.0",
+    "open": "^6.4.0",
     "readline": "^1.3.0"
   },
   "devDependencies": {
+    "ava": "^2.4.0",
     "babel-eslint": "^10.0.3",
     "eslint": "^6.5.1",
-    "eslint-plugin-babel": "^5.2.1"
+    "eslint-plugin-babel": "^5.2.1",
+    "sinon": "^7.5.0"
   }
 }
