+++ yo-dawg.js
@@ -1,59 +1,3 @@
-const fetch = require('node-fetch');
-const opn = require('opn');
-const consoleReader = require('./utils/console-reader');
+const yoDawg = require('./yo-dawg/');
 
-const pimpMyMeme = async () => {
-	console.log('\nXzibit needs some details...\n');
-
-	// Set some defaults.
-	let captions = {
-		one: 'cars',
-		two: ['car', 'car'],
-		three: ['drive', 'drive']
-	};
-
-	const caption1 = await consoleReader.question('Yo dawg, we heard you like _____ : ');
-	if (caption1 !== '') {
-		captions.one = caption1;
-	}
-
-	const caption2 = await consoleReader.question('So we put a ____ in your ____ : (e.g. - first, second) ');
-	if (caption2 !== '') {
-		captions.two = caption2.split(',');
-	}
-
-	const caption3 = await consoleReader.question('so you can ____ while you ____ : (e.g. - first, second) ');
-	if (caption3 !== '') {
-		captions.three = caption3.split(',');
-	}
-
-	generateMeme(captions);
-};
-
-const generateMeme = ({ one, two, three }) => {
-
-    let topText = `Yo dawg, we heard you like ${one}`;
-    let bottomText = `So we put a ${two[0]} in your ${two[1]} so you can ${three[0]} while you ${three[1]}`
-
-    let url = 'https://api.imgflip.com/caption_image?template_id=101716'
-    + '&text0=' + encodeURI(topText)
-    + '&text1=' + encodeURI(bottomText)
-    + '&username=dankestofmemes2000'
-    + '&password=bR[DVqvjPGd87rAUVR8XvBzkNK@viC8W';
-    
-	fetch( url, {
-        method: "POST"
-	})
-	.then(res => { 
-        return res.json(); })
-	.then(r => {
-		return r.data.url;
-	})
-	.then(url => {
-        opn(url);
-        process.exit();
-    })
-    .catch(console.error);  
-};
-
-pimpMyMeme();
+yoDawg();
