+++ yo-dawg/index.js
@@ -1,0 +1,3 @@
+const yoDawg = require('./yo-dawg');
+
+module.exports = yoDawg.pimpMyMeme;
